Separate platform specific code
* Move all platform specific code into properly guarded files.
* Fix tags all across the board.
* Restrict tests and benchmarks to supported CPUs.
* Remove unexported duplicated functions.

Package now builds on all platforms, but a lot will of course remain unsupported, but exported functions are properly returning errors and doesn't call platform specific code.

Replaces #10
