+++ finalize_structurals_amd64.s
@@ -1,4 +1,4 @@
-//+build !noasm !appengine
+//+build !noasm !appengine gc
 // AUTO-GENERATED BY C2GOASM -- DO NOT EDIT
 
 TEXT ·_finalize_structurals(SB), $0-48
