+++ find_quote_mask_and_bits_amd64.s
@@ -1,4 +1,4 @@
-//+build !noasm !appengine
+//+build !noasm !appengine gc
 // AUTO-GENERATED BY C2GOASM -- DO NOT EDIT
 
 DATA LCDATA1<>+0x000(SB)/8, $0x2222222222222222
