+++ find_structural_bits_amd64.s
@@ -1,4 +1,4 @@
-//+build !noasm !appengine
+//+build !noasm !appengine gc
 
 TEXT ·_find_structural_bits(SB), $0-72
