+++ parse_string_amd64.go
@@ -1,5 +1,6 @@
 //+build !noasm
 //+build !appengine
+//+build gc
 
 /*
  * MinIO Cloud Storage, (C) 2020 MinIO, Inc.
