+++ simdjson_amd64.go
@@ -1,3 +1,7 @@
+//+build !appengine
+//+build !noasm
+//+build gc
+
 /*
  * MinIO Cloud Storage, (C) 2020 MinIO, Inc.
  *
