+++ stage1_find_marks_amd64_test.go
@@ -1,3 +1,7 @@
+//+build !noasm
+//+build !appengine
+//+build gc
+
 /*
  * MinIO Cloud Storage, (C) 2020 MinIO, Inc.
  *
@@ -21,8 +25,6 @@ import (
 	"strings"
 	"testing"
 )
-
-const demo_json = `{"Image":{"Width":800,"Height":600,"Title":"View from 15th Floor","Thumbnail":{"Url":"http://www.example.com/image/481989943","Height":125,"Width":100},"Animated":false,"IDs":[116,943,234,38793]}}`
 
 func reverseBinary(input string) string {
 	// Get Unicode code points.
